+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added alert for unclosed tickets on the submission page
 - fixed mass messaging not showing dig units as source
 - fixed haul truck dispatch logs not showing all changes
+- fixed time allocation editor not rendering for time_allocation_report shortcut
 
 ## 0.10.1
 - fixed ticket modal width